Manage Quotes
Two new pages are created when you install the Manage Quotes module, QUOH adn QUOT. The individual pages have been added as well as a framework package which will update the default pages to match the code and style of the Shadows ReadyTheme. If you are added this to a different ReadyTheme, you may need to modify the code to match.